Model Problems

The following model problems are available and are dispatched via the following types.

ExtendableASGFEM.PoissonProblemPrimal — Type

abstract type PoissonProblemPrimal <: ExtendableASGFEM.AbstractModelProblem

Poisson problem with linear stochastic coefficient $a$ that seeks $u$ such that

$-\mathrm{div}(a(y,x) \nabla u(y,x)) = f(x) \quad \text{for } (y,x) \in \Gamma \times D$

ExtendableASGFEM.LogTransformedPoissonProblemPrimal — Type

abstract type LogTransformedPoissonProblemPrimal <: ExtendableASGFEM.AbstractModelProblem

Poisson problem with exponential stochastic coefficient $e^a$ that seeks $u$ such that

$-\mathrm{div}(e^a(y,x) \nabla u(y,x)) = f(x) \quad \text{for } (y,x) \in \Gamma \times D$

The stochastic Galerkin FEM solves the equivalent transformed problem

$-\mathrm{div}(\nabla u(y,x)) - \nabla a(y,x) \cdot \nabla u(y,x) = e^{-a(y,x)} f(x) \quad \text{for } (y,x) \in \Gamma \times D$

ExtendableASGFEM.LogTransformedPoissonProblemDual — Type

abstract type LogTransformedPoissonProblemDual <: ExtendableASGFEM.AbstractModelProblem

Dual log-transformed formulation of the Poisson problem with exponential stochastic coefficient (WIP).