ONBasis

ExtendableASGFEM.ONBasis — Type

ONBasis(
    OBT::Type{<:OrthogonalPolynomialType},
    maxorder;
    ...
) -> ONBasis{Float64}
ONBasis(
    OBT::Type{<:OrthogonalPolynomialType},
    maxorder,
    maxquadorder;
    T
) -> ONBasis{Float64}

Constructs an ONBasis for the given OrthogonalPolynomialType (and the associated weight function) and the maximal polynomial order.

ExtendableASGFEM.ONBasis — Type

struct ONBasis{T<:Real, OBT<:OrthogonalPolynomialType, npoly, nquad}

Structure that stores information of an orthogonal basis, i.e., their norms, quadrature rules and values at reference points, and the result of the last evaluate call.

ExtendableASGFEM.evaluate — Method

evaluate(
    ONB::ONBasis{T, OBT, maxorder},
    x;
    normalize
) -> Vector

Evaluates all basis polynomials at point x

ExtendableASGFEM.integral — Method

integral(ONB::ONBasis, j; normalize) -> Any

Evaluates the integral of a single basis function.

ExtendableASGFEM.scalar_product — Method

scalar_product(ONB::ONBasis, j, k; normalize) -> Any

Evaluates the scalar product of two basis functions.

ExtendableASGFEM.triple_product — Method

triple_product(ONB::ONBasis, j, k, l; normalize) -> Any

Evaluates the triple product of three basis functions.

ExtendableASGFEM.triple_product_y — Method

triple_product_y(ONB::ONBasis, j, k; normalize) -> Any

Evaluates the triple product of two basis functions and y.