SGFEVector

ExtendableASGFEM.SGFEVector — Type

struct SGFEVector{T, Tv, Ti, ONBType<:ONBasis, MIType}

structure that builds upon ExtendableFEMBase.FEVector and adds information on the stochastic discretization, i.e. the used TensoriedBasis

ExtendableASGFEM.SGFEVector — Method

SGFEVector(
    FES::Array{<:ExtendableFEMBase.FESpace{Tv, Ti}, 1},
    TB::TensorizedBasis{Tv, ONBType, MIType};
    active_modes,
    T,
    unames
) -> SGFEVector

contructs an SGFEVector for the given (spatial) finite element space FES and the (tensorized basis of the stochastic discretization) TB.

Base.getindex — Method

getindex(
    SGFEV::SGFEVector,
    u::Int64,
    i::Int64
) -> ExtendableFEMBase.FEVectorBlock

returns the FEVectorBlock for the i-th stochastic mode of the u-th unknown

Base.getindex — Method

getindex(
    SGFEV::SGFEVector,
    i::Int64
) -> ExtendableFEMBase.FEVectorBlock

returns the i-th stochastic mode

Base.length — Method

length(SGFEV::SGFEVector) -> Int64

returns the length of the full vector, i.e., the total number of degrees of freedom

Base.show — Method

show(io::IO, SGFEV::SGFEVector)

Custom show function for FEVector that prints some information on its blocks.

Base.size — Method

size(SGFEV::SGFEVector)

returns a tuple with the number of active stochastic modes and the number of spatial degrees of freedom

ExtendableASGFEM.fetypes — Method

fetypes(SGFEV::SGFEVector) -> Vector

returns the finite element types of the (spatial) FE spaces

ExtendableASGFEM.num_multiindices — Method

num_multiindices(SGFEV::SGFEVector) -> Int64

returns the number of active modes (that are used from the stored tensorized basis)

ExtendableASGFEM.set_sample! — Method

set_sample!(SGFEV::SGFEVector, S::AbstractVector)

evaluates the SGVector at that sample and stores the results in SGFEV.FEV.